Validate haproxy stat areas. Fixes #16411

+++ net/pfSense-pkg-haproxy-devel/files/usr/local/www/haproxy/haproxy_stats.php
@@ -123,19 +123,19 @@ haproxy_display_top_tabs_active($haproxy_tab_array['haproxy'], "stats");
 
 	<?php
 if (isset($_GET['showstatresolvers'])){
-	$showstatresolversname = $_GET['showstatresolvers'];
 	echo "<td colspan='2'>";
-	echo "Resolver statistics: $sticktablename<br/>";
-	$res = haproxy_socket_command("show resolvers $showstatresolversname");
+	echo "Resolver statistics:<br/>";
+	$res = haproxy_socket_command("show resolvers globalresolvers");
 	foreach($res as $line){
 		echo "<br/>".print_r($line,true);
 	}
 	echo "</td>";
-} elseif (isset($_GET['showsticktablecontent'])){
+} elseif (isset($_GET['showsticktablecontent']) &&
+	 (array_key_exists($_GET['showsticktablecontent'], haproxy_get_tables()))) {
 	$sticktablename = $_GET['showsticktablecontent'];
 	echo "<td colspan='2'>";
-	echo "Contents of the sticktable: $sticktablename<br/>";
-	$res = haproxy_socket_command("show table $sticktablename");
+	echo "Contents of the sticktable: " . htmlspecialchars($sticktablename) . "<br/>";
+	$res = haproxy_socket_command("show table {$sticktablename}");
 	foreach($res as $line){
 		echo "<br/>".print_r($line,true);
 	}
